% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmnetCoxTools.R
\name{Cox_bootstrapping}
\alias{Cox_bootstrapping}
\title{Cox glmnet model bootstrapping for survival association hypothesis testing}
\usage{
Cox_bootstrapping(
  X_data,
  y_data,
  alpha = 0,
  lambda = c(exp(seq(-4, 6, 0.1))),
  pre.CV = TRUE,
  lambda_opt = 0,
  free_cores = 2,
  iter = 200,
  log_AUC = c(1:2),
  Patient.Z = c(1:2),
  Drug.Z = c(1:2),
  RCPC = c(0, 1, 2, 3, 4)
)
}
\arguments{
\item{X_data}{Input drug screen data}

\item{y_data}{Survival data}

\item{alpha}{Penalty type. 0 for L2 (Ridge), 1 for L1 (Lasso), or any number between 0 and 1 for elastic net penalty.}

\item{lambda}{Vector of regularization penalties if pre.CV = FALSE.}

\item{pre.CV}{Logical. pre.CV=TRUE if optimal lambda is provided.}

\item{lambda_opt}{Optimal lambda if pre.CV=TRUE.}

\item{free_cores}{Number of free cores.}

\item{iter}{Number of test iterations.}

\item{log_AUC}{log2 transform of AUC. 1 for yes, 2 for no.}

\item{Patient.Z}{Patient-wise standardization. 1 for yes, 2 for no.}

\item{Drug.Z}{Drug-wise standardization. 1 for yes, 2 for no.}

\item{RCPC}{Removal of confounding principal components.}

\item{test.n}{Vector of test set balance. First integer is the number of survivals, and second integer is the number of deaths.}
}
\value{
A list containing coefficients and associated statistics.
}
\description{
Survival association statistics.
}
\author{
Aram N. Andersen \email{aram.n.andersen@gmail.com}
}
