% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmnetCoxTools.R
\name{Cox_forecasting_drug_withdrawal_2}
\alias{Cox_forecasting_drug_withdrawal_2}
\title{Cox forecasting glmnet, drug withdrawal testing}
\usage{
Cox_forecasting_drug_withdrawal_2(
  X_data,
  y_data,
  Reduce = c("Iterative", "Iterative2"),
  alpha = 0,
  lambda = c(exp(seq(-4, 6, 0.1))),
  free_cores = 2,
  test.n = c(6, 4),
  iter = 200,
  log_AUC = c(1:2),
  Patient.Z = c(1:2),
  Drug.Z = c(1:2),
  RCPC = c(0, 1, 2, 3, 4),
  path = path
)
}
\arguments{
\item{X_data}{Input drug screen data}

\item{y_data}{Survival data}

\item{Reduce}{Variable removal for dimension reduction. "Iterative" will reduce the model by removing one and one variable starting from highest C-index gain until optimal model is found, while re-testing variables as the model is reduced.}

\item{alpha}{Penalty type. 0 for L2 (Ridge), 1 for L1 (Lasso), or any number between 0 and 1 for elastic net penalty.}

\item{lambda}{Vector of regularization penalties.}

\item{free_cores}{Number of free cores.}

\item{test.n}{Vector of test set balance. First integer is the number of survivals, and second integer is the number of deaths.}

\item{iter}{Number of test iterations.}

\item{log_AUC}{log2 transform of AUC. 1 for yes, 2 for no.}

\item{Patient.Z}{Patient-wise standardization. 1 for yes, 2 for no.}

\item{Drug.Z}{Drug-wise standardization. 1 for yes, 2 for no.}

\item{RCPC}{Removal of confounding principal components.}
}
\value{
A list containing C-index test results from the variable withdrawal routine.
Contains the test results from the parent data, and the C-index differential upon drug withdrawal.
Optimization results from removing non-predictive variables.
}
\description{
C-index variable importance testing of Cox partial likelihood glmnet model for survival forecasting from ex vivo drug screens.
}
\author{
Aram N. Andersen \email{aram.n.andersen@gmail.com}
}
